Little improvements and 'add' method

+++ prefix-search/stringcoding/fc.go
@@ -5,7 +5,7 @@ type FC struct {
 }
 
 func NewFC(strings []string) FC {
-	return FC{New(strings, func (prefixLen, stringLen uint) uint {
+	return FC{New(strings, func (prefixLen, stringLen uint64) uint64 {
 		return prefixLen
 	})}
 }
